+++ src/main/java/ru/clevertec/ecl/config/SpringWebAppConfig.java
@@ -2,13 +2,8 @@ package ru.clevertec.ecl.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
-import org.springframework.beans.factory.config.BeanFactoryPostProcessor;
-import org.springframework.beans.factory.config.YamlPropertiesFactoryBean;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
-import org.springframework.core.io.ClassPathResource;
 import org.springframework.http.converter.HttpMessageConverter;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
@@ -17,8 +12,6 @@ import ru.clevertec.ecl.util.LocalDateTimeSerializer;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Objects;
-import java.util.Properties;
 
 @Configuration
 @EnableWebMvc
@@ -37,15 +30,5 @@ public class SpringWebAppConfig implements WebMvcConfigurer {
         SimpleModule simpleModule = new SimpleModule().addSerializer(LocalDateTime.class, new LocalDateTimeSerializer());
         objectMapper.registerModule(simpleModule);
         return objectMapper;
-    }
-
-    @Bean
-    public BeanFactoryPostProcessor beanFactoryPostProcessor() {
-        PropertySourcesPlaceholderConfigurer propertyConfigurer = new PropertySourcesPlaceholderConfigurer();
-        YamlPropertiesFactoryBean yaml = new YamlPropertiesFactoryBean();
-        yaml.setResources(new ClassPathResource("application.yml"));
-        Properties yamlObject = Objects.requireNonNull(yaml.getObject(), "Could not load yml");
-        propertyConfigurer.setProperties(yamlObject);
-        return propertyConfigurer;
     }
 }
