+++ src/main/resources/application-dev.yml
@@ -1,9 +1,8 @@
 spring:
   datasource:
-    url: jdbc:postgres://localhost:5432/clevertec
+    url: jdbc:postgresql://localhost:5432/clevertec
     username: postgres
     password: 172143
   hibernate:
-    dialect: org.hibernate.dialect.PostgreSQL10Dialect
-    connection:
-      driver: org.postgresql.Driver
+    format_sql: true
+    show_sql: true
