+++ src/main/resources/application-prod.yml
@@ -1,0 +1,5 @@
+spring:
+  datasource:
+    url:
+    username:
+    password:
