+++ src/main/resources/application.yml
@@ -1,9 +1,9 @@
 spring:
-  datasource:
-    url: jdbc:postgresql://localhost:5432/clevertec
-    username: postgres
-    password: 172143
   hibernate:
     dialect: org.hibernate.dialect.PostgreSQL10Dialect
     connection:
       driver: org.postgresql.Driver
+    format_sql: false
+    show_sql: false
+    hikari:
+      maximumPoolSize: 8
